refactor: move INSTANCE_OFFSET to mmdet.structures
Relocate the `INSTANCE_OFFSET` constant from
`mmdet.evaluation.functional.panoptic_utils` to `mmdet.structures`
so it can be shared without importing from an evaluation module into
model code.

- Define `INSTANCE_OFFSET = 1000` in `mmdet/structures/__init__.py`
  and export it via `__all__`
- Replace the inline definition in `panoptic_utils.py` with a
  re-export (`from mmdet.structures import INSTANCE_OFFSET`)
- Update imports in `heuristic_fusion_head.py` and
  `maskformer_fusion_head.py` to use the new canonical location
