+++ mmdet/models/seg_heads/panoptic_fusion_heads/heuristic_fusion_head.py
@@ -5,8 +5,8 @@ import torch
 from mmengine.structures import InstanceData, PixelData
 from torch import Tensor
 
-from mmdet.evaluation.functional import INSTANCE_OFFSET
 from mmdet.registry import MODELS
+from mmdet.structures import INSTANCE_OFFSET
 from mmdet.utils import InstanceList, OptConfigType, OptMultiConfig, PixelList
 from .base_panoptic_fusion_head import BasePanopticFusionHead
