+++ mmdet/models/seg_heads/panoptic_fusion_heads/maskformer_fusion_head.py
@@ -6,9 +6,8 @@ import torch.nn.functional as F
 from mmengine.structures import InstanceData, PixelData
 from torch import Tensor
 
-from mmdet.evaluation.functional import INSTANCE_OFFSET
 from mmdet.registry import MODELS
-from mmdet.structures import SampleList
+from mmdet.structures import INSTANCE_OFFSET, SampleList
 from mmdet.structures.mask import mask2bbox
 from mmdet.utils import OptConfigType, OptMultiConfig
 from .base_panoptic_fusion_head import BasePanopticFusionHead
